Move page configuration to main js folder and shorten its name
